Major Changes
This commit changes content on all of the pages with the exception of the Education page. I have added a theme bar to the bottom of the homepage which will eventually allow the user to change themes. I have also added to my about bio and included links to various projects on the skills page.